Extract fighter/shuttle FSMs out of ModelLoader.js
ModelLoader.js was 750 lines, ~580 of them two unrelated state machines
(NPC fighter AI: form/attack/evade/returnToAttack, and shuttle AI:
depart/return/slowDown/rise/docked/lower/turnToDepart) that had nothing
to do with loading a GLTF model.

Pulled them into webGL/js/ai/FighterFSM.js and ShuttleFSM.js as factory
functions that build and return a FiniteStateMachine instance (the
existing generic engine in utils/FiniteStateMachine.js is unchanged).
ModelLoader.js now just loads the model, wires up the animation mixer,
and calls the right factory - down to ~170 lines.

Pure extraction, no behavior change: same states, same transitions,
same NpcControls calls.
